pipe(shareReplay(1))

+++ src/app/view-product/view-product.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
 import { ProductService } from '../product.service';
 
 @Component({
@@ -8,14 +9,15 @@ import { ProductService } from '../product.service';
 })
 export class ViewProductComponent implements OnInit {
    productList:any="";
-  constructor(private product:ProductService) {
+  constructor(private product:ProductService,private router:Router) {
     this.product.viewProduct().subscribe(result=>{
       this.productList=result;
     })
    }
 
    deleteProductHtml(PId:any,SNo:any){
-     confirm("Are your sure");
+     let t=confirm("Are your sure");
+     if(t)
      this.product.deleteProduct(PId).subscribe(result=>{
        if(result){
         this.productList.splice(SNo,1);
@@ -26,7 +28,10 @@ export class ViewProductComponent implements OnInit {
       //  this.productList=result;
      })
    }
-
+   
+   editProductHtml(pid:any){
+    this.router.navigate(['edit-product',pid]);
+   }
   ngOnInit(): void {
   }
